Skip long tags on build tags-indexes. #752

+++ plugins/tags/plugin.cpp
@@ -207,10 +207,10 @@ namespace golos { namespace plugins { namespace tags {
     ) {
         config_file_options.add_options()
             (
-                "tags-number", boost::program_options::value<uint16_t>()->default_value(5),
+                "tags-number", boost::program_options::value<uint8_t>()->default_value(5),
                 "Maximum number of tags"
             ) (
-                "tag-max-length", boost::program_options::value<uint32_t>()->default_value(512),
+                "tag-max-length", boost::program_options::value<uint16_t>()->default_value(512),
                 "Maximum length of tag"
             );
     }
